# docs/api-spec.openapi.yaml
openapi: 3.1.0
info:
  title: Nashid Platform API
  description: High-performance streaming, synchronized lyrics, and vocabulary learning API.
  version: 1.0.0
servers:
  - url: https://api.nashid.app/v1
    description: Production Server
  - url: http://localhost:4000/v1
    description: Local Development Server

paths:
  /nasheeds:
    get:
      summary: List all published nasheeds
      description: Returns a list of nasheeds with pagination and metadata.
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: tag
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of nasheeds
          content:
            application/json:
              schema:
                type: object
                properties:
                  nasheeds:
                    type: array
                    items:
                      $ref: '#/components/schemas/NasheedSummary'
                  nextCursor:
                    type: string
                    nullable: true

  /nasheeds/{id}:
    get:
      summary: Get full nasheed details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Full nasheed object
          content:
            application/json:
              schema:
                type: object
                properties:
                  nasheed:
                    $ref: '#/components/schemas/NasheedDetail'

  /nasheeds/first:
    get:
      summary: Get default or featured initial nasheed
      responses:
        '200':
          description: First nasheed ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string

  /nasheeds/by-ids:
    post:
      summary: Fetch multiple nasheeds by their IDs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: List of nasheeds
          content:
            application/json:
              schema:
                type: object
                properties:
                  nasheeds:
                    type: array
                    items:
                      $ref: '#/components/schemas/NasheedSummary'

  /nasheeds/{id}/play:
    post:
      summary: Record play event for analytics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Play recorded

  /nasheeds/{id}/favorite:
    post:
      summary: Toggle favorite status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isFavorite:
                  type: boolean
      responses:
        '200':
          description: Favorite status updated

  /dictionary/by-timed-word-ids:
    post:
      summary: Get dictionary details for word IDs in timed lyrics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                timedWordIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Dictionary word map
          content:
            application/json:
              schema:
                type: object
                properties:
                  dictionary:
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/DictionaryWord'

  /vocabulary/words:
    get:
      summary: Get user's saved vocabulary words
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Saved vocabulary list
          content:
            application/json:
              schema:
                type: object
                properties:
                  words:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavedWord'
    post:
      summary: Save a new word to vocabulary for review
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                wordId:
                  type: string
                lemmaKey:
                  type: string
      responses:
        '201':
          description: Word saved to vocabulary

  /playlists:
    get:
      summary: Get user playlists
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Playlists list
          content:
            application/json:
              schema:
                type: object
                properties:
                  playlists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Playlist'
    post:
      summary: Create new playlist
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                isPublic:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Playlist created

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    NasheedSummary:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        artistName:
          type: string
        coverUrl:
          type: string
        audioUrl:
          type: string
        lyricsJsonUrl:
          type: string
        durationMs:
          type: integer
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        totalWords:
          type: integer
        newWordsCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        playCount:
          type: integer
        likeCount:
          type: integer
        savedCount:
          type: integer

    NasheedDetail:
      allOf:
        - $ref: '#/components/schemas/NasheedSummary'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            waveformUrl:
              type: string
              nullable: true
            isFavorite:
              type: boolean
            subtitleLanguages:
              type: array
              items:
                type: string

    DictionaryWord:
      type: object
      properties:
        id:
          type: string
        arabic:
          type: string
        normalized:
          type: string
        partOfSpeech:
          type: string
        meanings:
          type: array
          items:
            type: string
        meaningsEn:
          type: array
          items:
            type: string
        root:
          type: string
          nullable: true
        plural:
          type: string
          nullable: true
        wazn:
          type: string
          nullable: true
        bab:
          type: string
          nullable: true
        present:
          type: string
          nullable: true
        imperative:
          type: string
          nullable: true
        example:
          type: string
          nullable: true

    SavedWord:
      type: object
      properties:
        id:
          type: string
        word:
          $ref: '#/components/schemas/DictionaryWord'
        masteryLevel:
          type: integer
          default: 0
        nextReviewAt:
          type: string
          format: date-time

    Playlist:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        itemCount:
          type: integer
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time